backend/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
cached-property==1.5.2
certifi==2024.2.2
click==8.1.7
DateTime==5.5
distro==1.9.0
exceptiongroup==1.2.1
fastapi==0.103.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.7
importlib-metadata==6.7.0
openai==1.25.0
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==0.21.1
python-multipart==0.0.8
pytz==2024.1
sniffio==1.3.1
starlette==0.27.0
tqdm==4.66.2
typing_extensions==4.7.1
uuid==1.30
uvicorn==0.22.0
zipp==3.15.0
zope.interface==6.4.post2
